+++ tutorials/sources/Pods/PlayerCore/PlayerCore/components/Ad.swift
@@ -2,6 +2,10 @@
 //  Licensed under the terms of the MIT License. See LICENSE.md file in project root for terms.
 import Foundation
 
+public enum AdType {
+    case preroll, midroll
+}
+
 public struct Ad {
     public let playedAds: Set<UUID>
     public let midrolls: [Midroll]
@@ -42,14 +46,6 @@ func reduce(state: Ad, action: Action) -> Ad {
     }
     
     switch action {
-    case let action as AdRequest:
-        return Ad(playedAds: state.playedAds,
-                  midrolls: state.midrolls,
-                  mp4AdCreative: nil,
-                  vpaidAdCreative: nil,
-                  currentAd: state.currentAd,
-                  currentType: action.type)
-        
     case let action as VRMCore.AdRequest:
         return Ad(playedAds: state.playedAds,
                   midrolls: state.midrolls,
@@ -57,28 +53,6 @@ func reduce(state: Ad, action: Action) -> Ad {
                   vpaidAdCreative: nil,
                   currentAd: state.currentAd,
                   currentType: action.type)
-        
-    case let action as ShowAd:
-        var playedAds = state.playedAds
-        playedAds.insert(action.id)
-        switch action.creative {
-        case .mp4(let creatives):
-            return Ad(playedAds: playedAds,
-                      midrolls: state.midrolls,
-                      mp4AdCreative: creatives.first,
-                      vpaidAdCreative: nil,
-                      currentAd: .play,
-                      currentType: state.currentType)
-        case .vpaid(let creatives):
-            return Ad(playedAds: playedAds,
-                      midrolls: state.midrolls,
-                      mp4AdCreative: nil,
-                      vpaidAdCreative: creatives.first,
-                      currentAd: .play,
-                      currentType: state.currentType)
-        case .none:
-            fatalError("AdCreative.none has to create SkipAd action")
-        }
         
     case let action as ShowMP4Ad:
         var playedAds = state.playedAds
@@ -115,12 +89,12 @@ func reduce(state: Ad, action: Action) -> Ad {
     case is ShowContent,
          is SkipAd,
          is AdPlaybackFailed,
-         is AdError,
+         is VPAIDActions.AdError,
          is AdStartTimeout,
          is AdMaxShowTimeout,
-         is AdStopped,
-         is AdSkipped,
-         is AdNotSupported:
+         is VPAIDActions.AdStopped,
+         is VPAIDActions.AdSkipped,
+         is VPAIDActions.AdNotSupported:
         return Ad(playedAds: state.playedAds,
                   midrolls: state.midrolls,
                   mp4AdCreative: state.mp4AdCreative,