+++ tutorials/sources/Pods/PlayerCore/PlayerCore/components/Mute.swift
@@ -13,10 +13,10 @@ func reduce(state: Mute, action: Action) -> Mute {
         return Mute(player: true, vpaid: true)
     case is PlayerUnmute:
         return Mute(player: false, vpaid: false)
-    case let action as AdVolumeChange:
+    case let action as VPAIDActions.AdVolumeChange:
         let isVPAIDMuted = action.volume == 0
         return Mute(player: state.player, vpaid: isVPAIDMuted)
-    case is ShowMP4Ad, is ShowVPAIDAd, is ShowAd:
+    case is ShowMP4Ad, is ShowVPAIDAd:
         return Mute(player: state.player, vpaid: state.player)
     default: return state        
     }